token_problem/requirements.txt
appdirs==1.4.4
asgiref==3.3.0
black==20.8b1
click==7.1.2
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
flake8==3.8.4
flake8-black==0.2.1
gunicorn==20.0.4
install==1.3.4
Markdown==3.3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.4
regex==2020.10.28
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
